requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
dj-static==0.0.6
Django==4.2.6
django-stdimage==6.0.1
gunicorn==21.2.0
idna==3.4
linkpreview==0.8.1
lxml==4.9.3
packaging==23.2
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
static3==0.7.0
tzdata==2023.3
urllib3==1.26.7
